all: refactor package structure and naming
