+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.msfjarvis.todo">
+    package="dev.msfjarvis.lobsters">
 
     <application
         android:name=".Application"
@@ -9,11 +9,11 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.ComposeToDo">
+        android:theme="@style/Theme.Lobsters">
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
-            android:theme="@style/Theme.ComposeToDo.NoActionBar">
+            android:theme="@style/Theme.Lobsters.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
