+++ app/src/main/java/dev/msfjarvis/lobsters/urllauncher/UrlLauncher.kt
@@ -1,4 +1,4 @@
-package dev.msfjarvis.todo.urllauncher
+package dev.msfjarvis.lobsters.urllauncher
 
 interface UrlLauncher {
   fun launch(url: String)
