+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Compose-ToDo</string>
+    <string name="app_name">lobste.rs</string>
 </resources>
