+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.ComposeToDo" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.Lobsters" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/white</item>
         <item name="colorPrimaryVariant">@color/white</item>
@@ -15,12 +15,12 @@
         <!-- Customize your theme here. -->
     </style>
 
-    <style name="Theme.ComposeToDo.NoActionBar">
+    <style name="Theme.Lobsters.NoActionBar">
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
     </style>
 
-    <style name="Theme.ComposeToDo.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+    <style name="Theme.Lobsters.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
 
-    <style name="Theme.ComposeToDo.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
+    <style name="Theme.Lobsters.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
 </resources>
