+++ settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = "Compose-ToDo"
+rootProject.name = "lobste.rs"
 include ':app'
 include ':data'
